+++ requests_cache/models/response.py
@@ -160,8 +160,8 @@ class CachedResponse(RichMixin, BaseResponse):
         """Returns a PreparedRequest for the next request in a redirect chain, if there is one."""
         return self._next.prepare() if self._next else None
 
-    def reset_expiration(self, expire_after: ExpirationTime) -> bool:
-        """Set a new expiration for this response, and determine if it is now expired"""
+    def reset_expiration(self, expire_after: ExpirationTime):
+        """Set a new expiration for this response"""
         self.expires = get_expiration_datetime(expire_after)
         return self.is_expired